* moved HashExtension to specfic class   * ActiveResource fixed
